create and read part of CRUD also dependencies like laravel form and datatable is intrigated

+++ database/migrations/2019_04_19_085005_create_clients_table.php
@@ -23,7 +23,7 @@ class CreateClientsTable extends Migration
             $table->timestamp('dob');
             $table->string('mode_of_contact');
             $table->string('email')->unique();
-            $table->smallInteger('phone')->unique();
+            $table->bigInteger('phone')->unique();
             $table->rememberToken();
             $table->timestamps();
         });
